+++ docs/CHANGELOG.md
@@ -1,3 +1,33 @@
+## 1.0.3
+Released on 19th June 2020
+
+### Added
+
+#### Staking
+
+* GetLastBlockFees Handler at /api/staking/lastblockfees
+* GetAddressFromPublicKey Handler at /api/staking/publickeytoaddress
+
+#### Consensus:
+
+* GetStatus Handler at /api/consensus/status
+* GetGenesisDocument Handler at /api/consensus/genesisdocument
+
+#### Registry:
+
+* GetNodeStatus Handler at /api/registry/nodestatus
+* GetRegistryEvents Handler at /api/registry/events
+
+### Updated
+
+#### Staking
+
+* GetAccounts at /api/staking/accounts has been changed to GetAddresses at /api/staking/addresses
+* GetAccountInfo at /api/staking/accountinfo has been changed to GetAccount at /api/staking/account, and "ownerKey" query parameter has been changed to "address"
+* GetDelegations "ownerKey" query parameter has been changed to "address"
+* GetDebondingDelegations "ownerKey" query parameter has been changed to "address"
+
+
 ## 1.0.2
 
 Released on May 2020